refactor: introduce useHandleSettingScrollable (review point 6)
